requirements.txt
# =============================================================================
# Dark Factory v2 — Automated Financial Compliance & SEC Filing Synthesis
# =============================================================================
# Core Dependencies
# =============================================================================
openai>=1.12.0
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.2.0
aiohttp>=3.9.3
httpx>=0.27.0
pyyaml>=6.0.1
jinja2>=3.1.3
rich>=13.7.0
python-frontmatter>=1.1.0

# =============================================================================
# Testing & Benchmarking
# =============================================================================
pytest>=8.0.0
pytest-asyncio>=0.23.0
flask>=3.0.0

# =============================================================================
# Framework Dependencies (install from source / private registry)
# -----------------------------------------------------------------------------
# These are not available on public PyPI. Install per their documentation:
#
#   Hermes Agent — Persistent-memory agent orchestration framework
#     pip install git+https://github.com/hermes-agent/hermes-agent.git
#
#   Parchi — Relay Daemon and CLI browser endpoints
#     pip install git+https://github.com/parchi-project/parchi.git
#
#   WebArena-Infinity — Web agent evaluation framework
#     pip install git+https://github.com/web-arena-x/webarena-infinity.git
# =============================================================================
